refactor : 구조 및 파일명 변경
